requirements.txt
djangorestframework
django
open3d
markitdown
pdf2image
ExtractTable